requirements.txt
ipython==8.12.3
langchain==0.3.17
langchain_chroma==0.2.1
langchain_core==0.3.34
langchain_google_vertexai==2.0.13
langchain_openai==0.3.4
langgraph==0.2.70
langsmith==0.3.6
pydantic==2.10.6
python-dotenv==1.0.1
python-telegram-bot==21.10
rich==13.9.4
